requirements.txt
# Core Python dependencies for Proxy-GS.
# Install PyTorch, torch-scatter, and the local CUDA extensions separately
# following the instructions in README.md.

colorama==0.4.6
einops==0.8.1
jaxtyping==0.3.3
lpips==0.1.4
numpy==2.3.3
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast.git@abb07ca0358f3a21c3942b50c54aa1eacd329af9
open3d==0.19.0
opencv-python-headless==4.10.0.84
OpenEXR==3.4.4
Pillow==11.3.0
plyfile==1.1.3
scipy==1.16.3
tqdm==4.67.1
trimesh==4.9.0
wandb==0.23.0
